# Add .NET 5 target and improve WriteString performance with SIMD