@jtattermusch 

Optimization improves writing larger all-ASCII strings.
* 105 and 10080 bytes are much faster
* 10 is on the border, but shows a small improvement
* 1 and 4 use existing path

Before

```

|                   Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |------------- |------------ |-----------:|----------:|----------:|------:|------:|------:|----------:|
| WriteString_WriteContext |        10080 |           1 | 146.551 us | 1.7018 us | 1.5918 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |           4 |  48.298 us | 0.4151 us | 0.3467 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |          10 |  24.715 us | 0.1720 us | 0.1436 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |         105 |  11.238 us | 0.0745 us | 0.0697 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |       10080 |   8.917 us | 0.1159 us | 0.1084 us |     - |     - |     - |         - |

```

Process 4 chars at a time

```
|                   Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |------------- |------------ |-----------:|----------:|----------:|------:|------:|------:|----------:|
| WriteString_WriteContext |        10080 |           1 | 148.329 us | 0.8029 us | 0.6705 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |           4 |  48.822 us | 0.7867 us | 0.6974 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |          10 |  21.464 us | 0.1850 us | 0.1730 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |         105 |   5.907 us | 0.0802 us | 0.0711 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |       10080 |   4.195 us | 0.0380 us | 0.0356 us |     - |     - |     - |         - |

```

Process 4 chars at a time + SSE2/ARM SIMD

```
|                   Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |------------- |------------ |-----------:|----------:|----------:|------:|------:|------:|----------:|
| WriteString_WriteContext |        10080 |           1 | 171.521 us | 1.9836 us | 1.8555 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |           4 |  53.462 us | 0.5362 us | 0.4753 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |          10 |  21.630 us | 0.1186 us | 0.1109 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |         105 |   5.100 us | 0.0815 us | 0.0722 us |     - |     - |     - |         - |
| WriteString_WriteContext |        10080 |       10080 |   3.385 us | 0.0192 us | 0.0170 us |     - |     - |     - |         - |