@jtattermusch 2021-01-28
The other writeString optimization PR has been merged, let's rebase?

The C# build is currently failing:
```
A compatible installed .NET Core SDK for global.json version [5.0.100] from [/tmp/protobuf/protobuf/global.json] was not found
Install the [5.0.100] .NET Core SDK or update [/tmp/protobuf/protobuf/global.json] with an installed .NET Core SDK:
```
A new enough version of the runtime needs to be installed in the test scripts or in the dockerfile.