@JamesNK 2021-01-28
Updated:

Before:
```
|                           Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |------------- |------------ |-----------:|----------:|----------:|------:|------:|------:|----------:|
|         WriteString_WriteContext |        10080 |           1 | 195.968 us | 2.4081 us | 2.2525 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |           4 |  50.370 us | 0.5141 us | 0.4293 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |           4 |  61.447 us | 1.1956 us | 1.2793 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |          10 |  20.653 us | 0.0704 us | 0.0658 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |          10 |  26.459 us | 0.4852 us | 0.4301 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |         105 |  11.325 us | 0.1994 us | 0.1866 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |         105 |   8.223 us | 0.0481 us | 0.0450 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |       10080 |   8.838 us | 0.0754 us | 0.0705 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |       10080 |   4.413 us | 0.0647 us | 0.0605 us |     - |     - |     - |         - |
```

After:
```
|                           Method | BytesToWrite | encodedSize |       Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |------------- |------------ |-----------:|----------:|----------:|------:|------:|------:|----------:|
|         WriteString_WriteContext |        10080 |           1 | 192.867 us | 1.5161 us | 1.3440 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |           4 |  50.275 us | 0.9840 us | 0.8723 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |           4 |  64.725 us | 0.4342 us | 0.3849 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |          10 |  20.809 us | 0.1590 us | 0.1410 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |          10 |  26.623 us | 0.2650 us | 0.2478 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |         105 |   4.780 us | 0.0446 us | 0.0417 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |         105 |   9.668 us | 0.0416 us | 0.0369 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |       10080 |   2.919 us | 0.0184 us | 0.0172 us |     - |     - |     - |         - |
| WriteNonAsciiString_WriteContext |        10080 |       10080 |   4.838 us | 0.0304 us | 0.0269 us |     - |     - |     - |         - |
```

Tests that use the new path:

Before:
```
|         WriteString_WriteContext |        10080 |         105 |  11.325 us | 0.1994 us | 0.1866 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |       10080 |   8.838 us | 0.0754 us | 0.0705 us |     - |     - |     - |         - |
```

After:
```
|         WriteString_WriteContext |        10080 |         105 |   4.780 us | 0.0446 us | 0.0417 us |     - |     - |     - |         - |
|         WriteString_WriteContext |        10080 |       10080 |   2.919 us | 0.0184 us | 0.0172 us |     - |     - |     - |         - |
```

🔥