@JamesNK 2021-01-28
> A new enough version of the runtime needs to be installed in the test scripts or in the dockerfile.

I've attempted to fix this but can't trigger the build to test. If it doesn't work can you finish making the changes required here.